[rtloader] renaming cgo_free.c/h sources to memory.c/h
