+++ rtloader/common/builtins/datadog_agent.c
@@ -3,9 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019 Datadog, Inc.
 #include "datadog_agent.h"
-#include "cgo_free.h"
-
-#include <stringutils.h>
+#include "memory.h"
+#include "stringutils.h"
 
 // these must be set by the Agent
 static cb_get_version_t cb_get_version = NULL;
