+++ rtloader/common/builtins/tagger.c
@@ -4,9 +4,9 @@
 // (https://www.datadoghq.com/).
 // Copyright 2019 Datadog, Inc.
 
-#include "cgo_free.h"
+#include "memory.h"
 #include "stringutils.h"
-#include <tagger.h>
+#include "tagger.h"
 
 // these must be set by the Agent
 static cb_tags_t cb_tags = NULL;
