+++ rtloader/common/memory.h
