+++ rtloader/test/common/cgo_free_two.go
@@ -6,7 +6,7 @@ package testcommon
 #cgo CFLAGS: -I../../common
 #cgo !windows LDFLAGS: -L../../two/ -ldatadog-agent-two
 #cgo windows LDFLAGS: -L../../two/ -ldatadog-agent-two.dll
-#include <cgo_free.h>
+#include <memory.h>
 
 void c_callCgoFree(void *ptr) {
 	cgo_free(ptr);
